requirements.txt
argon2-cffi==21.3.0
asyncpg==0.26.0
cryptography==37.0.4
sanic==22.6.2
sanic-jwt==1.8.0
tortoise-orm==0.19.2
pycryptodome==3.15.0